+++ src/DependencyInjection/MonterHealthApiFilterExtension.php
@@ -27,9 +27,9 @@ class MonterHealthApiFilterExtension extends Extension
             $container->setAlias('monter_health_api_filter.parameter_collection_factory', $config['parameter_collection_factory']);
         }
 
-        // handle option to override the default Annotation Reader
-        if(null !== $config['annotation_reader']) {
-            $container->setAlias('monter_health_api_filter.annotation_reader', $config['annotation_reader']);
+        // handle option to override the default Attribute Reader
+        if(null !== $config['attribute_reader']) {
+            $container->setAlias('monter_health_api_filter.attribute_reader', $config['attribute_reader']);
         }
 
         // handle option to override the default ApiFilterFactory
