+++ src/Filter/SearchFilter.php
@@ -2,7 +2,7 @@
 
 namespace MonterHealth\ApiFilterBundle\Filter;
 
-use MonterHealth\ApiFilterBundle\Annotation\ApiFilter;
+use MonterHealth\ApiFilterBundle\Attribute\ApiFilter;
 use MonterHealth\ApiFilterBundle\Parameter\Collection;
 use MonterHealth\ApiFilterBundle\Parameter\Command;
 use MonterHealth\ApiFilterBundle\Util\QueryNameGeneratorInterface;
